refactor(backend): isolate generated image attempt persistence
